Add a note about Fluentd Enterprise Edition
This is one of those cases where a long commit message is
warranted because I have some explaining to do.

As many of you know, Treasure Data is a big sponsor of Fluentd,
and by popular demand, they began a subscription business to help
companies scale their Fluentd deployment.

I was recently at Elasticon in SF, and I realized that

1. Fluentd has a lot of users
2. Very few of them knows about Treasure Data offering support for it

Furthermore, some users had chosen to move away from Fluentd
because they were not sure if they could get enterprise-grade support.

As a passionate contributor of Fluentd, Treasure Data cares about
Fluentd greatly, and I hope folks don't mind a bit of commercial
advertisement in the form on HelloBar.

As for the choice of HelloBar, It's largely because my CSS skill is
non-exitent. I am happy to accept a PR that substitutes it with some
combination of HTML/CSS/JavaScript.