Switch from yui-compressor to sassc,uglifier
As yui-compressor is not maintained anymore [1], so
it should migrate from it.

The past PR [2] tried to migrate to sass and uglifier, but now sass is
also deprecated. so sassc and uglifier are the reasonable choice for
migration.

[1] https://rubygems.org/gems/yui-compressor
[2] https://github.com/fluent/fluentd-website/pull/78/commits/71f175814fd902fd5d4c2dc151b0e66e2d84c540
[3] https://sass-lang.com/ruby-sass

In the previous migration PR, https://github.com/fluent/fluentd-website/pull/125
set :assets_css_compressor should be :sass instead of :sassc.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>